setup.sh: don't install software automatically
Installing software is complicated, and I think it is reasonable to
ask the user to install some software on their own. We could extend
the error messages with suggestions on how to find the packages, or
links to the install pages for docker and docker-compose.

That said, I'm not sure 82 lines of fairly complex code to cover just
2 distros is worth it.